backend/requirements.txt
fastapi
uvicorn
langgraph
langchain
langchain-groq
langchain-community
sqlalchemy
pandas
matplotlib
reportlab
python-dotenv
pydantic
requests
python-multipart
